fix: various test failures

+++ LibAtem/Serialization/Int32Attribute.cs
@@ -72,7 +72,7 @@ namespace LibAtem.Serialization
 
         public override object GetRandom(Random random)
         {
-            int range = ScaledMax - ScaledMin;
+            double range = (double)ScaledMax - (double)ScaledMin;
             return (random.NextDouble() * range + ScaledMin) / Scale;
         }
 
@@ -174,7 +174,7 @@ namespace LibAtem.Serialization
 
     public class UInt32DScaleAttribute : UInt32DAttribute
     {
-        public UInt32DScaleAttribute(double max = 1) : base(uint.MaxValue / max, 0, (uint) (uint.MaxValue / max))
+        public UInt32DScaleAttribute(double max = 1) : base(uint.MaxValue / max, 0, (uint)(uint.MaxValue / max))
         {
         }
     }
